+++ easygui/boxes/utils.py
